backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
passlib[bcrypt]
sqlalchemy
python-jose
gunicorn